Move UnmarkedWsNoticeDetails into features
